Miscellaneous test framework fixes

+++ ci-scripts/test/pingdirectory/01-ldap-tests.sh
@@ -3,9 +3,15 @@
 SCRIPT_HOME=$(cd $(dirname ${0}); pwd)
 . "${SCRIPT_HOME}"/../../common.sh "${1}"
 
+if skipTest "${0}"; then
+  log "Skipping test ${0}"
+  exit 0
+fi
+
 if "${DEV_TEST_ENV}" = 'true'; then
   if pgrep -f docker > /dev/null; then
     docker run --rm pingidentity/ldap-sdk-tools ldapsearch \
+      --terse \
       --hostname "${PINGDIRECTORY_ADMIN}" \
       --port 636 \
       --bindDN 'cn=administrator' \
@@ -15,10 +21,11 @@ if "${DEV_TEST_ENV}" = 'true'; then
       --baseDN "cn=config" \
       --searchScope base "(&)" 1.1
   else
-    echo 'Docker daemon required to run this test in dev environments'
+    log 'Docker daemon required to run this test in dev environments'
   fi
 else
   /opt/tools/ldapsearch \
+    --terse \
     --hostname "${PINGDIRECTORY_ADMIN}" \
     --port 636 \
     --bindDN 'cn=administrator' \
